+++ src/Config/Input/F-14BU/joystick/Throttle - HOTAS Warthog F-14BU.diff.lua
@@ -6,6 +6,14 @@ local diff = {
 		["a2004cdnil"] = { ["name"] = "Throttle Both", ["removed"] = { [1] = { ["key"] = "JOY_Z" } } },
 	},
 	["keyDiffs"] = {
+		["dnilp3020u3020cd18vdnilvp-1vu0"] = {
+			["added"] = { [1] = { ["key"] = "JOY_BTN7" } },
+			["name"] = "Speed brake retract",
+		},
+		["dnilp3020u3020cd18vdnilvp1vu0"] = {
+			["added"] = { [1] = { ["key"] = "JOY_BTN8" } },
+			["name"] = "Speed brake extend",
+		},
 		["d3024pnilu3024cd17vd1vpnilvu0"] = {
 			["added"] = { [1] = { ["key"] = "JOY_BTN11" } },
 			["removed"] = { [1] = { ["key"] = "JOY_BTN3" } },